refactor: change file database entirely and change a little bit server file
