+++ db/runner.go
@@ -1,0 +1,57 @@
+package db
+
+import (
+	"fmt"
+	"log"
+	entity_account "transaction/module/account/entity"
+	entity_category "transaction/module/category/entity"
+	entity_status "transaction/module/status/entity"
+	entity_transaction "transaction/module/transaction/entity"
+	entity_user "transaction/module/user/entity"
+
+	config "transaction/configs"
+
+	"gorm.io/gorm"
+)
+
+var db *gorm.DB
+
+func GetGormDB() *gorm.DB {
+	return db
+}
+
+func StartDatabase() {
+	databaseConfiguration := createDatabaseStringConfig()
+	mysql := NewMysql(databaseConfiguration)
+
+	if err := mysql.connect(); err != nil {
+		log.Fatal(err)
+	}
+
+	loadMigrations(mysql.db)
+
+	db = mysql.db
+}
+
+func createDatabaseStringConfig() string {
+	databaseStringConfig := fmt.Sprintf(
+		"%s:%s@tcp(%s:%s)/%s%s",
+
+		config.Mysql.USER,
+		config.Mysql.PASSWORD,
+		config.Mysql.HOST,
+		config.Mysql.PORT,
+		config.Mysql.DB,
+		config.Mysql.ADDITIONAL_CONFIGS,
+	)
+
+	return databaseStringConfig
+}
+
+func loadMigrations(db *gorm.DB) {
+	db.AutoMigrate(&entity_account.Account{})
+	db.AutoMigrate(&entity_category.Category{})
+	db.AutoMigrate(&entity_status.Status{})
+	db.AutoMigrate(&entity_transaction.Transaction{})
+	db.AutoMigrate(&entity_user.User{})
+}
