+++ docker-compose.yml
@@ -3,9 +3,9 @@ services:
   super-app-db:
     image: mysql:5.7
     environment:
-      MYSQL_DATABASE: "${DATABASE}"
-      MYSQL_ROOT_PASSWORD: '${DB_PASSWORD}'
+      MYSQL_DATABASE: "${MYSQL_DATABASE}"
+      MYSQL_ROOT_PASSWORD: '${MYSQL_PASSWORD}'
     ports:
-      - '${DB_PORT}:${DB_PORT}'
+      - '${MYSQL_PORT}:${MYSQL_PORT}'
     expose:
-      - '${DB_PORT}'
+      - '${MYSQL_PORT}'
